web-scraper-with-ai-2-main/requirements.txt
beautifulsoup4==4.12.2
cohere==4.5.0
deep-translator==1.8.4
Flask==2.3.2
httpx==0.24.0
requests==2.31.0
selenium==4.11.2
urllib3==1.26.16
webdriver-manager
requests
networkx
transformers
matplotlib
plotly
folium
IPython
spacy
azure-core
azure-ai-textanalytics
openpyxl
pandas
pdfplumber
xlrd
PyPDF2
docx2txt
